# cmd/api/internal/swaggerdoc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is all apis for the trip-itinary application.",
        "title": "Trip-Itinary API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/day_sessions/{id}/plan-versions": {
            "get": {
                "description": "Create planversions for a daysessions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plan"
                ],
                "summary": "Create Plan Version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "daysession ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/plans_application_dto.PlanVersionDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create planversions for a daysessions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plan"
                ],
                "summary": "Create Plan Version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "daysession ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "daysession ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/plans_application_dto.PlanVersionDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_module_plans_infrastructure_driving_rest.apiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        },
                        "requestID": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "plans_application_dto.PlanVersionDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "daysessionID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    }
}